Separate Web & Native Link Activity Contract
# Conflicts:
#	paymentsheet/src/test/java/com/stripe/android/link/LinkActivityContractTest.kt
