+++ paymentsheet/src/test/java/com/stripe/android/link/LinkActivityContractTest.kt
@@ -1,24 +1,23 @@
 package com.stripe.android.link
 
-import androidx.core.os.BundleCompat
+import android.content.Intent
+import android.net.Uri
 import androidx.test.core.app.ApplicationProvider
 import com.google.common.truth.Truth.assertThat
-import com.stripe.android.PaymentConfiguration
 import com.stripe.android.core.utils.FeatureFlags
-import com.stripe.android.networking.StripeRepository
 import com.stripe.android.testing.FeatureFlagTestRule
-import org.junit.After
-import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.mock
-import org.mockito.kotlin.any
+import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
 class LinkActivityContractTest {
+
+    private val args = LinkActivityContract.Args(TestFactory.LINK_CONFIGURATION)
 
     @get:Rule
     val featureFlagTestRule = FeatureFlagTestRule(
@@ -26,54 +25,81 @@ class LinkActivityContractTest {
         isEnabled = false
     )
 
-    @Before
-    fun before() {
-        PaymentConfiguration.init(
-            context = ApplicationProvider.getApplicationContext(),
-            publishableKey = "pk_test_abcdefg",
-        )
-    }
-
-    @After
-    fun after() {
-        PaymentConfiguration.clearInstance()
-    }
-
     @Test
     fun `LinkActivityContract creates intent with URL with native link disabled`() {
         featureFlagTestRule.setEnabled(false)
 
+        val expectedIntent = Intent()
+        val webLinkActivityContract = mock<WebLinkActivityContract>()
+        whenever(webLinkActivityContract.createIntent(ApplicationProvider.getApplicationContext(), args))
+            .thenReturn(expectedIntent)
+
+        val contract = linkActivityContract(webLinkActivityContract = webLinkActivityContract)
+
+        val actualIntent = contract.createIntent(ApplicationProvider.getApplicationContext(), args)
+
+        assertThat(expectedIntent).isEqualTo(actualIntent)
+    }
+
+    @Test
+    fun `LinkActivityContract parses result with webLinkActivityContract`() {
+        featureFlagTestRule.setEnabled(false)
+
+        val expectedIntent = Intent()
         val args = LinkActivityContract.Args(TestFactory.LINK_CONFIGURATION)
-        val stripeRepository = mock<StripeRepository>()
-        whenever(stripeRepository.buildPaymentUserAgent(any())).thenReturn("test")
-        val contract = LinkActivityContract(stripeRepository)
+        val webLinkActivityContract = mock<WebLinkActivityContract>()
+        whenever(webLinkActivityContract.createIntent(ApplicationProvider.getApplicationContext(), args))
+            .thenReturn(expectedIntent)
+
+        val contract = linkActivityContract(webLinkActivityContract = webLinkActivityContract)
+
         val intent = contract.createIntent(ApplicationProvider.getApplicationContext(), args)
-        assertThat(intent.component?.className).isEqualTo(LinkForegroundActivity::class.java.name)
-        assertThat(intent.extras?.getString(LinkForegroundActivity.EXTRA_POPUP_URL)).startsWith(
-            "https://checkout.link.com/#"
-        )
+        intent.data = Uri.EMPTY
+
+        contract.parseResult(0, intent)
+
+        verify(webLinkActivityContract).parseResult(0, intent)
     }
 
     @Test
     fun `LinkActivityContract creates intent with with NativeLinkArgs when native link is enabled`() {
         featureFlagTestRule.setEnabled(true)
 
-        val args = LinkActivityContract.Args(TestFactory.LINK_CONFIGURATION)
-        val stripeRepository = mock<StripeRepository>()
-        whenever(stripeRepository.buildPaymentUserAgent(any())).thenReturn("test")
-        val contract = LinkActivityContract(stripeRepository)
-        val intent = contract.createIntent(ApplicationProvider.getApplicationContext(), args)
-        assertThat(intent.component?.className).isEqualTo(LinkActivity::class.java.name)
+        val expectedIntent = Intent()
+        val nativeLinkActivityContract = mock<NativeLinkActivityContract>()
+        whenever(nativeLinkActivityContract.createIntent(ApplicationProvider.getApplicationContext(), args))
+            .thenReturn(expectedIntent)
 
-        val actualArg = intent.extras?.let {
-            BundleCompat.getParcelable(it, LinkActivity.EXTRA_ARGS, NativeLinkArgs::class.java)
-        }
-        assertThat(actualArg).isEqualTo(
-            NativeLinkArgs(
-                configuration = TestFactory.LINK_CONFIGURATION,
-                publishableKey = "pk_test_abcdefg",
-                stripeAccountId = null
-            )
+        val contract = linkActivityContract(nativeLinkActivityContract = nativeLinkActivityContract)
+
+        val actualIntent = contract.createIntent(ApplicationProvider.getApplicationContext(), args)
+
+        assertThat(expectedIntent).isEqualTo(actualIntent)
+    }
+
+    @Test
+    fun `LinkActivityContract parses result with nativeLinkActivityContract`() {
+        featureFlagTestRule.setEnabled(true)
+
+        val nativeLinkActivityContract = mock<NativeLinkActivityContract>()
+        whenever(nativeLinkActivityContract.createIntent(ApplicationProvider.getApplicationContext(), args))
+            .thenReturn(Intent())
+
+        val contract = linkActivityContract(nativeLinkActivityContract = nativeLinkActivityContract)
+
+        val intent = contract.createIntent(ApplicationProvider.getApplicationContext(), args)
+        contract.parseResult(0, intent)
+
+        verify(nativeLinkActivityContract).parseResult(0, intent)
+    }
+
+    private fun linkActivityContract(
+        webLinkActivityContract: WebLinkActivityContract = mock(),
+        nativeLinkActivityContract: NativeLinkActivityContract = mock()
+    ): LinkActivityContract {
+        return LinkActivityContract(
+            nativeLinkActivityContract = nativeLinkActivityContract,
+            webLinkActivityContract = webLinkActivityContract
         )
     }
 }
